requirements.txt
discord.py
google-genai
groq
aiohttp
yt-dlp
tiktoken
Pillow
python-dotenv
requests
youtube-transcript-api